trading-bot/requirements.txt
alpaca-py==0.15.0
pandas==2.1.4
numpy==1.24.3
asyncio-mqtt==0.11.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
yfinance==0.2.18
ta==0.10.2